requirements.txt
streamlit
langchain-community
langchain-openai
faiss-cpu
openai
python-dotenv
tiktoken
